+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+### Added
+- Introduced `message_simple()` to show inline feedback message anywhere on the site
+
 ### Changed
 - Improved translation of
     - Messages